subsystem: TLBusWrapper base class has all special coupling methods
