+++ src/main/scala/subsystem/PeripheryBus.scala
@@ -37,9 +37,7 @@ case class PeripheryBusParams(
 
 class PeripheryBus(params: PeripheryBusParams, name: String)(implicit p: Parameters)
     extends TLBusWrapper(params, name)
-    with CanHaveBuiltInDevices
-    with CanAttachTLSlaves {
-
+{
   private val fixer = LazyModule(new TLFIFOFixer(TLFIFOFixer.all))
   private val node: TLNode = params.atomics.map { pa =>
     val in_xbar = LazyModule(new TLXbar)
@@ -60,12 +58,4 @@ class PeripheryBus(params: PeripheryBusParams, name: String)(implicit p: Paramet
   def busView: TLEdge = fixer.node.edges.in.head
 
   val builtInDevices: BuiltInDevices = BuiltInDevices.attach(params, outwardNode)
-
-  def toTile
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => TLInwardNode): NoHandle = {
-    to("tile" named name) { FlipRendering { implicit p =>
-      gen :*= TLWidthWidget(params.beatBytes) :*= TLBuffer(buffer) :*= outwardNode
-    }}
-  }
 }
